rafactor: refactoring the two classes of exception into just one
