+++ src/main/java/com/maisprati/forum/exception/ConnectionException.java
@@ -1,4 +1,4 @@
-package com.maisprati.forum.exceptions;
+package com.maisprati.forum.exception;
 
 public class ConnectionException extends RuntimeException {
     public ConnectionException(String message) {
