--- src/main/java/com/maisprati/forum/exceptions/UserNotFoundException.java
@@ -1,7 +1,0 @@
-package com.maisprati.forum.exceptions;
-
-public class UserNotFoundException extends RuntimeException {
-    public UserNotFoundException(String message) {
-        super(message);
-    }
-}
